+++ src/queue.c
@@ -261,7 +261,7 @@ dump_block_len_queue(block_len_queue_t* queue, int fd)
 }
 
 block_len_queue_t*
-read_block_len_queue(void* input_map, int offset, int end)
+read_block_len_queue(void* input_map, long offset, long end)
 {
     if(input_map == NULL)
         error("read_block_len_queue: input_map is NULL");
